fix: fix gremlin detection and make it boring
KoL now sends the monster ID, so we know whether this gremlin has a
tool. This also means that the old code does not even work anymore,
because the gremlins with tools are not handled at all.
We no longer need to detect the tool gremlins ourselves, so greatly
simplify the gremlin handler (and make it much more boring).